+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Changes:
 - Autofocus on new tab input
 - Add tab when pressing enter instead of click
 - Fix tab removal not saving
+- (Broken) Emote menu for 7tv emotes